Add lazy loading to twig extension 'media_library_widget'
Add lazy loading to twig extension 'media_library_widget'

Optimizing page load speed by lazy loading media_library_widget.
