+++ src/Frontend/Modules/MediaLibrary/Twig/Extensions/FrontendHelperExtensions.php
@@ -2,53 +2,20 @@
 
 namespace Frontend\Modules\MediaLibrary\Twig\Extensions;
 
-use Frontend\Modules\MediaLibrary\Helper\FrontendHelper;
-use Twig_Markup;
+use Frontend\Modules\MediaLibrary\Twig\AppRuntime;
+use Twig_Extension;
+use Twig_SimpleFunction;
 
-class FrontendHelperExtensions extends \Twig_Extension
+class FrontendHelperExtensions extends Twig_Extension
 {
-    /**
-     * @var FrontendHelper
-     */
-    private $frontendHelper;
-
-    public function __construct(FrontendHelper $frontendHelper)
-    {
-        $this->frontendHelper = $frontendHelper;
-    }
-
     public function getFunctions(): array
     {
         return [
-            new \Twig_SimpleFunction(
+            new Twig_SimpleFunction(
                 'media_library_widget',
-                [$this, 'parseWidget'],
+                [AppRuntime::class, 'parseWidget'],
                 ['is_safe' => ['html']]
             ),
         ];
-    }
-
-    /**
-     * Parses the widget
-     *
-     * @param string $mediaWidgetAction - The ClassName from the Media widget you want to use.
-     * @param string $mediaGroupId - The MediaGroup id you want to parse
-     * @param string $title - You can give your optional custom title.
-     * @param string $module - You can parse a widget from a custom module. Default is the "MediaLibrary" module.
-     *
-     * @return string
-     */
-    public function parseWidget(
-        string $mediaWidgetAction,
-        string $mediaGroupId,
-        string $title = null,
-        string $module = null
-    ): string {
-        return $this->frontendHelper->parseWidget(
-            $mediaWidgetAction,
-            $mediaGroupId,
-            $title,
-            $module
-        );
     }
 }
